+++ app/_components/sidenav.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import NavLinks from '@/components/nav-links';
+import NavLinks from '../_components/nav-links';
 import { PowerIcon } from '@heroicons/react/24/outline';
 import { signOut } from '@/auth';
