+++ app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/login-form';
+import LoginForm from '../_components/login-form';
 import { Suspense } from 'react';
 
 export default function Page(){
